+++ src/main/java/com/binance/connector/client/WebSocketApiClient.java
@@ -5,11 +5,15 @@ import com.binance.connector.client.impl.websocketapi.WebSocketApiGeneral;
 import com.binance.connector.client.impl.websocketapi.WebSocketApiMarket;
 import com.binance.connector.client.impl.websocketapi.WebSocketApiTrade;
 import com.binance.connector.client.impl.websocketapi.WebSocketApiUserDataStream;
-import com.binance.connector.client.utils.WebSocketCallback;
+import com.binance.connector.client.utils.websocketcallback.WebSocketClosedCallback;
+import com.binance.connector.client.utils.websocketcallback.WebSocketClosingCallback;
+import com.binance.connector.client.utils.websocketcallback.WebSocketFailureCallback;
+import com.binance.connector.client.utils.websocketcallback.WebSocketMessageCallback;
+import com.binance.connector.client.utils.websocketcallback.WebSocketOpenCallback;
 
 public interface WebSocketApiClient {
-    void connect(WebSocketCallback onMessageCallback);
-    void connect(WebSocketCallback onOpenCallback, WebSocketCallback onMessageCallback, WebSocketCallback onClosingCallback, WebSocketCallback onFailureCallback);
+    void connect(WebSocketMessageCallback onMessageCallback);
+    void connect(WebSocketOpenCallback onOpenCallback, WebSocketMessageCallback onMessageCallback, WebSocketClosingCallback onClosingCallback, WebSocketClosedCallback onClosedCallback, WebSocketFailureCallback onFailureCallback);
     void close();
     WebSocketApiGeneral general();
     WebSocketApiMarket market();