+++ src/main/java/com/binance/connector/client/utils/websocketcallback/WebSocketMessageCallback.java
@@ -1,4 +1,4 @@
-package com.binance.connector.client.utils;
+package com.binance.connector.client.utils.websocketcallback;
 
 @FunctionalInterface
 public interface WebSocketMessageCallback {