+++ src/test/java/unit/MockData.java
@@ -3,9 +3,13 @@ package unit;
 import java.util.LinkedHashMap;
 
 import com.binance.connector.client.utils.RequestBuilder;
-import com.binance.connector.client.utils.WebSocketCallback;
 import com.binance.connector.client.utils.WebSocketConnection;
 import com.binance.connector.client.utils.signaturegenerator.HmacSignatureGenerator;
+import com.binance.connector.client.utils.websocketcallback.WebSocketClosedCallback;
+import com.binance.connector.client.utils.websocketcallback.WebSocketClosingCallback;
+import com.binance.connector.client.utils.websocketcallback.WebSocketFailureCallback;
+import com.binance.connector.client.utils.websocketcallback.WebSocketMessageCallback;
+import com.binance.connector.client.utils.websocketcallback.WebSocketOpenCallback;
 
 import okhttp3.Request;
 
@@ -23,10 +27,14 @@ public final class MockData {
             put("key2", intValue2);
         }};
     public static final HmacSignatureGenerator HMAC_SIGNATURE_GENERATOR = new HmacSignatureGenerator(SECRET_KEY);
-    public static final WebSocketCallback noopCallback = msg -> { };
+    public static final WebSocketOpenCallback noopOpenCallback = msg -> { };
+    public static final WebSocketMessageCallback noopMessageCallback = msg -> { };
+    public static final WebSocketClosingCallback noopClosingCallback = (code, reason) -> { };
+    public static final WebSocketClosedCallback noopClosedCallback = (code, reason) -> { };
+    public static final WebSocketFailureCallback noopFailureCallback = (throwable, response) -> { };
     public static final String WS_BASE_URL = "wss://base.url";
     public static final Request WS_REQUEST = RequestBuilder.buildWebSocketRequest(WS_BASE_URL);
-    public static final WebSocketConnection WS_CONNECTION = new WebSocketConnection(noopCallback, noopCallback, noopCallback, noopCallback, WS_REQUEST, null);
+    public static final WebSocketConnection WS_CONNECTION = new WebSocketConnection(noopOpenCallback, noopMessageCallback, noopClosingCallback, noopClosedCallback, noopFailureCallback, WS_REQUEST, null);
     public static final String WS_ID = "websocketId";
     private MockData() {
     }